+++ Changelog.md
@@ -7,6 +7,7 @@
 * Support for synchronizations
 * Navigation withing collections (first, next and previous links)
 * Better type handling when deserializing JSON (e.g: DateTime instead of strings)
+* Fetch relationships
 
 ## 0.1.0